ativ-01

+++ ativ-00/src/Exercicio09.java
@@ -1,6 +1,6 @@
 public class Exercicio09 {
     public static void main(String[] args) throws Exception {
-        for (int i = 0; i <= 10; i++) {
+        for (byte i = 0; i <= 10; i++) {
             System.out.println(i + "x" + i + " = " + i*i);
         }
     }
